# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID gerado pelo cliente ao abrir o checkout. Repetir devolve o mesmo pedido.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária."
          },
          "403": {
            "description": "Papel sem permissão."
          },
          "409": {
            "description": "Loja fechada, inativa ou item indisponível."
          },
          "422": {
            "description": "Falha de validação."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária."
          },
          "403": {
            "description": "Papel sem permissão."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{orderId}": {
      "get": {
        "operationId": "OrdersController_find",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária."
          },
          "403": {
            "description": "Papel sem permissão."
          },
          "404": {
            "description": "Pedido não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{orderId}/cancellation": {
      "post": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária."
          },
          "403": {
            "description": "Papel sem permissão."
          },
          "404": {
            "description": "Pedido não encontrado."
          },
          "409": {
            "description": "O pedido não está mais aguardando a loja."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "PetDots API",
    "description": "Contrato canônico de fronteira do PetDots, gerado a partir dos schemas Zod.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "storeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "items": {
            "minItems": 1,
            "maxItems": 30,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "offerId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 20
                }
              },
              "required": [
                "offerId",
                "quantity"
              ]
            }
          }
        },
        "required": [
          "storeId",
          "items"
        ]
      },
      "OrderDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "code": {
            "type": "string",
            "description": "Seis caracteres legíveis — o número dito ao telefone."
          },
          "status": {
            "type": "string",
            "enum": [
              "PLACED",
              "ACCEPTED",
              "DISPATCHED",
              "DELIVERED",
              "REJECTED",
              "CANCELLED"
            ]
          },
          "store": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "slug": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "neighborhood": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "slug",
              "name",
              "neighborhood"
            ],
            "additionalProperties": false
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "productId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "productName": {
                  "type": "string",
                  "description": "Snapshot: o nome no momento da compra."
                },
                "productVariant": {
                  "type": "string"
                },
                "unitPriceCents": {
                  "type": "integer",
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991,
                  "minimum": 0
                },
                "quantity": {
                  "type": "integer",
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991,
                  "minimum": 0
                },
                "lineTotalCents": {
                  "type": "integer",
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991,
                  "minimum": 0
                },
                "fulfillment": {
                  "type": "string",
                  "enum": [
                    "FULFILLED",
                    "SUBSTITUTED",
                    "UNAVAILABLE"
                  ]
                }
              },
              "required": [
                "id",
                "productId",
                "productName",
                "productVariant",
                "unitPriceCents",
                "quantity",
                "lineTotalCents",
                "fulfillment"
              ],
              "additionalProperties": false
            }
          },
          "itemsTotalCents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "deliveryFeeCents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "serviceFeeCents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "totalCents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "deliveryAddress": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string"
              },
              "number": {
                "type": "string"
              },
              "complement": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "neighborhood": {
                "type": "string"
              },
              "postalCode": {
                "type": "string",
                "description": "Oito dígitos, sem hífen."
              },
              "reference": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "street",
              "number",
              "complement",
              "neighborhood",
              "postalCode",
              "reference"
            ],
            "additionalProperties": false
          },
          "contactName": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "placedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "acceptanceDeadlineAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "description": "Prazo de aceite, contado só em horário de funcionamento (ADR-0014 C2)."
          },
          "acceptedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          },
          "dispatchedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          },
          "rejectedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "enum": [
              "STORE_REJECTED",
              "ACCEPTANCE_EXPIRED",
              null
            ],
            "nullable": true
          },
          "cancellationReason": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "code",
          "status",
          "store",
          "items",
          "itemsTotalCents",
          "deliveryFeeCents",
          "serviceFeeCents",
          "totalCents",
          "deliveryAddress",
          "contactName",
          "contactPhone",
          "placedAt",
          "acceptanceDeadlineAt",
          "acceptedAt",
          "dispatchedAt",
          "deliveredAt",
          "cancelledAt",
          "rejectedAt",
          "rejectionReason",
          "cancellationReason",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
    }
  }
}